Added Wish Pack Sprite
Based on Jirachi's Tanzaku tags hanging on its head. "Candy" is written on it in Hisuian text. Hisuian is chosen because Sinnoh is the only region where you can reliably get Jirachi in without an event (BDSP) and Jirachi is very ancient (and also the sinnohian alphabet is not decoded yet LOL), and candy is chosen because in the Jirachi movie, the first wish Max makes is for snacks.