Updated Japanese Localization
By no means a perfect translation but should be serviceable. The old one hadn't been updated since before the item rework.
Excluding Togepi (I couldn't figure out how to do it) as well as a couple other odds and ends but all the main stuff is done.
As an aside, Zigzagoon and Linoone are missing from the English localization file (I added them to Japanese manually).